Fix check in HiveMetadata.getView and update its name
The method is supposed to get a Trino view that's going to be updated.
Its check for `translateHiveViews` was redundant, the calling code is
never supposed to operate on a Hive view, regardless of
`translateHiveViews` flag.

It was probably not user-facing problem as the calling code is likely
unreachable when Hive view translation is not enabled.